requirements.txt
Flask==3.0.0
dnspython==2.6.1
requests==2.31.0
speedtest-cli==2.1.3
statistics==1.0.3.5
logging==0.4.9.6